refactor: use u64 for DEFAULT_MAX_QUOTA_LIMIT

+++ syncserver/src/web/extractors/heartbeat_request.rs
@@ -13,7 +13,7 @@ use crate::{server::ServerState, web::error::ValidationErrorKind};
 #[derive(Clone, Copy, Debug, Serialize)]
 pub struct QuotaInfo {
     pub enabled: bool,
-    pub size: u32,
+    pub size: u64,
 }
 
 #[derive(Clone, Debug)]
